Move test_models into their own search sub modules
